scw-project/requirements.txt
djangorestframework
django
celery
psycopg2-binary